Fixing more tests
